Reworked sound recorder
