+++ intellectus-sound-recorder/SoundRecorders/ISoundSource.cs
@@ -1,0 +1,22 @@
+﻿using NAudio.Wave;
+using SoundRecorder.SoundListeners;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SoundRecorder.SoundRecorders
+{
+    public interface ISoundSource
+    {
+        WaveFormat GetWaveFormat();
+
+        void AddListener(ISoundListener listener);
+
+        void RemoveListener(ISoundListener listener);
+
+        void Start();
+
+        void Stop();
+
+    }
+}
